bunch of refactors to fix tests requiring transformers library
