+++ harness/determined/integrations/huggingface/__init__.py
@@ -1,4 +1,3 @@
 from determined.integrations.huggingface.hf_callback import (
     DetCallback,
-    get_hf_args_with_overwrites,
 )
