+++ harness/determined/pytorch/dsat/__init__.py
@@ -3,6 +3,7 @@ from determined.pytorch.dsat._utils import (
     get_ds_config_from_hparams,
     overwrite_deepspeed_config,
     get_ds_config_path_from_args,
+    get_hf_args_with_overwrites,
     replace_ds_config_file_using_overwrites,
 )
 from determined.pytorch.dsat._dsat_search_method import (
